feat: add abstract resolver
